requirements.txt
# Data Retrieval
yfinance>=0.2.30
yahooquery>=2.3.3
fredapi>=0.5.1
pandas_datareader>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0
certifi>=2023.11.17

# Data Processing and Analytics
numpy>=1.24.4
pandas>=2.0.3
scipy>=1.10.1
scikit-learn>=1.3.2
statsmodels>=0.14.0

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2
plotly>=5.18.0
streamlit>=1.22

# Jupyter and Interactive Development
jupyter>=1.0.0
notebook>=7.0.6
ipywidgets>=8.1.1

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.1
openpyxl>=3.1.2  # For Excel file support
tabulate>=0.9.0  # For pretty table printing

